rental类相关的service、dao以及简单的测试

+++ src/main/java/graduation/spring/erent/sys/base/BaseDao.java
@@ -6,10 +6,10 @@ public interface BaseDao<T> {
     void insert(T t);
     void update(T t);
     void delete(T t);
-    void delete(Long id);
+    void delete(int id);
     T find(T t);
     List<T> findList(T t);
     List<T> findAll();
-    T findById(Long id);
+    T findById(int id);
 
 }
