+++ src/main/java/graduation/spring/erent/sys/base/BaseModel.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class BaseModel {
-    protected Long id;
+    protected int id;
 }
